Folder refactor & very basic implementation of createManyImages
